Feature/flow weighted concentrations (#124)
* Rename surfacearea module to deposition_utils

The surfacearea module is exclusively a helper for the deposition module.
Rename it to deposition_utils to clarify this relationship and remove
its cross-references from concepts.rst.

* Add flow-weighted mean volume computation and corresponding tests for diffusion module

* Review flow-weighted concentration: remove dead code, add validation, tighten tests

Remove unused _erf_integral_time, _erf_integral_space_time, and
_erf_mean_space_time from diffusion.py along with their tests. Add
flow > 0 validation in diffusion infiltration_to_extraction to prevent
division by zero. Tighten test tolerances to machine precision, add
partially capped cell test, strengthen front-tracking constant flow
test to compare against advection reference, and rename misnamed
test_varying_flow_mass_conservation to test_constant_flow_mass_conservation.

* Use pytest -n auto for test commands

Update CLAUDE.md to run pytest with -n auto for tests/src, tests/examples, and tests/docs. This enables pytest-xdist parallel execution to speed up local and CI test runs.

* Add typing hints & ndarray conversions, fix tests

Add/strengthen type annotations (NDArray, Sequence, ArrayLike) across fronttracking and advection helpers and return signatures; replace many .values uses in tests with .to_numpy() to ensure explicit numpy arrays; add defensive casts (float/np.array) and dtype enforcement (sparse.eye dtype=float) to avoid type issues; guard ShockWave outlet crossing when velocity is None; add existence check for external 'uv' tool in test script; adjust one test tolerance and fix minor test/formatting indentation; add type: ignore comments where necessary to satisfy static typing. These changes aim to improve type-safety and runtime robustness between pandas/NumPy boundaries and to make tests more deterministic.

* Increase timeout for Python script execution in tests to prevent hanging

+++ src/gwtransport/fronttracking/plot.py
@@ -79,7 +79,7 @@ def plot_vt_diagram(
     if t_max is None:
         # Default to input data time range instead of simulation end time
         # Convert tedges[-1] from Timestamp to days from tedges[0]
-        t_max = (state.tedges[-1] - state.tedges[0]) / pd.Timedelta(days=1)
+        t_max = float((state.tedges[-1] - state.tedges[0]) / pd.Timedelta(days=1))
 
     if ax is None:
         _, ax = plt.subplots(figsize=figsize)
@@ -244,9 +244,9 @@ def plot_vt_diagram(
             # Fill between head and tail
             if valid_head and valid_tail and len(valid_head) == len(valid_tail):
                 ax.fill_between(
-                    t_h,
-                    v_h,
-                    v_t,
+                    t_h,  # type: ignore[possibly-undefined]
+                    v_h,  # type: ignore[possibly-undefined]
+                    v_t,  # type: ignore[possibly-undefined]
                     color="green",
                     alpha=0.1 if not wave.is_active else 0.2,
                 )
@@ -386,7 +386,7 @@ def plot_breakthrough_curve(
     if t_max is None:
         # Default to input data time range instead of simulation end time
         # Convert tedges[-1] from Timestamp to days from tedges[0]
-        t_max = (state.tedges[-1] - state.tedges[0]) / pd.Timedelta(days=1)
+        t_max = float((state.tedges[-1] - state.tedges[0]) / pd.Timedelta(days=1))
 
     # Use exact analytical segments
     segments = identify_outlet_segments(0.0, t_max, state.v_outlet, state.waves, state.sorption)
